+++ jeap-governance-domain/src/main/java/ch/admin/bit/jeap/governance/domain/scheduler/SchedulerRunRepository.java
@@ -1,0 +1,11 @@
+package ch.admin.bit.jeap.governance.domain.scheduler;
+
+import java.time.LocalDateTime;
+import java.util.Optional;
+
+public interface SchedulerRunRepository {
+
+    Optional<LocalDateTime> findLastRunDateTime(String jobName);
+
+    void saveLastRunDateTime(String jobName, LocalDateTime lastRunDateTime);
+}