+++ jeap-governance-persistence/src/main/resources/db/migration/V1_0_2__create-scheduler-run.sql
@@ -1,0 +1,6 @@
+-- scheduler_run
+CREATE TABLE scheduler_run
+(
+    job_name    VARCHAR   NOT NULL PRIMARY KEY,
+    last_run_at TIMESTAMP NOT NULL
+);